`delete` API実行時, 先に関連スケジュールを持つかどうか確認してから関連スケジュールを削除する処理を実施する仕様を追加

+++ backend/src/main/java/com/magu1436/chronolist/timeblocking/TimeBlockingController.java
@@ -254,9 +254,11 @@ public class TimeBlockingController {
     @DeleteMapping("timeBlock/delete")
     public ResponseEntity<Void> delete(@RequestBody TimeBlock timeBlock){
         if(ExistsTimeBlockById(timeBlock.getId())){
-            // relatedScheduleの削除
-            Integer scheduleId = timeBlock.getRelatedSchedule().getId();
-            schedulerMapper.deleteSchedule(scheduleId);
+            // TimeBlockのrelatedScheduleがNullでない場合はscheduleテーブルも削除する
+            if (timeBlock.getRelatedSchedule() != null){
+                Integer scheduleId = timeBlock.getRelatedSchedule().getId();
+                schedulerMapper.deleteSchedule(scheduleId);
+            }
             // TimeBlockの削除
             timeBlockMapper.deleteTimeBlock(timeBlock.getId());
             return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
